+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_api_impls.dart
@@ -244,6 +244,11 @@ class WebViewHostApiImpl extends WebViewHostApi {
       instanceManager.getInstanceId(client)!,
     );
   }
+
+  /// Helper method to convert instances ids to objects.
+  Future<void> setBackgroundColorFromInstance(WebView instance, int color) {
+    return setBackgroundColor(instanceManager.getInstanceId(instance)!, color);
+  }
 }
 
 /// Host api implementation for [WebSettings].
@@ -428,9 +433,13 @@ class JavaScriptChannelFlutterApiImpl extends JavaScriptChannelFlutterApi {
 
   @override
   void postMessage(int instanceId, String message) {
-    final JavaScriptChannel instance =
-        instanceManager.getInstance(instanceId) as JavaScriptChannel;
-    instance.postMessage(message);
+    final JavaScriptChannel? instance =
+        instanceManager.getInstance(instanceId) as JavaScriptChannel?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an JavaScriptChannel with instanceId: $instanceId',
+    );
+    instance!.postMessage(message);
   }
 }
 
@@ -473,22 +482,36 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
 
   @override
   void onPageFinished(int instanceId, int webViewInstanceId, String url) {
-    final WebViewClient instance =
-        instanceManager.getInstance(instanceId) as WebViewClient;
-    instance.onPageFinished(
-      instanceManager.getInstance(webViewInstanceId) as WebView,
-      url,
+    final WebViewClient? instance =
+        instanceManager.getInstance(instanceId) as WebViewClient?;
+    final WebView? webViewInstance =
+        instanceManager.getInstance(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
     );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+    );
+    instance!.onPageFinished(webViewInstance!, url);
   }
 
   @override
   void onPageStarted(int instanceId, int webViewInstanceId, String url) {
-    final WebViewClient instance =
-        instanceManager.getInstance(instanceId) as WebViewClient;
-    instance.onPageStarted(
-      instanceManager.getInstance(webViewInstanceId) as WebView,
-      url,
+    final WebViewClient? instance =
+        instanceManager.getInstance(instanceId) as WebViewClient?;
+    final WebView? webViewInstance =
+        instanceManager.getInstance(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
     );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+    );
+    instance!.onPageStarted(webViewInstance!, url);
   }
 
   @override
@@ -499,11 +522,21 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
     String description,
     String failingUrl,
   ) {
-    final WebViewClient instance =
-        instanceManager.getInstance(instanceId) as WebViewClient;
+    final WebViewClient? instance =
+        instanceManager.getInstance(instanceId) as WebViewClient?;
+    final WebView? webViewInstance =
+        instanceManager.getInstance(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+    );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+    );
     // ignore: deprecated_member_use_from_same_package
-    instance.onReceivedError(
-      instanceManager.getInstance(webViewInstanceId) as WebView,
+    instance!.onReceivedError(
+      webViewInstance!,
       errorCode,
       description,
       failingUrl,
@@ -517,10 +550,20 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
     WebResourceRequestData request,
     WebResourceErrorData error,
   ) {
-    final WebViewClient instance =
-        instanceManager.getInstance(instanceId) as WebViewClient;
-    instance.onReceivedRequestError(
-      instanceManager.getInstance(webViewInstanceId) as WebView,
+    final WebViewClient? instance =
+        instanceManager.getInstance(instanceId) as WebViewClient?;
+    final WebView? webViewInstance =
+        instanceManager.getInstance(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+    );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+    );
+    instance!.onReceivedRequestError(
+      webViewInstance!,
       _toWebResourceRequest(request),
       _toWebResourceError(error),
     );
@@ -532,12 +575,19 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
     int webViewInstanceId,
     WebResourceRequestData request,
   ) {
-    final WebViewClient instance =
-        instanceManager.getInstance(instanceId) as WebViewClient;
-    instance.requestLoading(
-      instanceManager.getInstance(webViewInstanceId) as WebView,
-      _toWebResourceRequest(request),
+    final WebViewClient? instance =
+        instanceManager.getInstance(instanceId) as WebViewClient?;
+    final WebView? webViewInstance =
+        instanceManager.getInstance(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
     );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+    );
+    instance!.requestLoading(webViewInstance!, _toWebResourceRequest(request));
   }
 
   @override
@@ -546,12 +596,19 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
     int webViewInstanceId,
     String url,
   ) {
-    final WebViewClient instance =
-        instanceManager.getInstance(instanceId) as WebViewClient;
-    instance.urlLoading(
-      instanceManager.getInstance(webViewInstanceId) as WebView,
-      url,
+    final WebViewClient? instance =
+        instanceManager.getInstance(instanceId) as WebViewClient?;
+    final WebView? webViewInstance =
+        instanceManager.getInstance(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
     );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+    );
+    instance!.urlLoading(webViewInstance!, url);
   }
 }
 
@@ -601,9 +658,13 @@ class DownloadListenerFlutterApiImpl extends DownloadListenerFlutterApi {
     String mimetype,
     int contentLength,
   ) {
-    final DownloadListener instance =
-        instanceManager.getInstance(instanceId) as DownloadListener;
-    instance.onDownloadStart(
+    final DownloadListener? instance =
+        instanceManager.getInstance(instanceId) as DownloadListener?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an DownloadListener with instanceId: $instanceId',
+    );
+    instance!.onDownloadStart(
       url,
       userAgent,
       contentDisposition,
@@ -655,11 +716,18 @@ class WebChromeClientFlutterApiImpl extends WebChromeClientFlutterApi {
 
   @override
   void onProgressChanged(int instanceId, int webViewInstanceId, int progress) {
-    final WebChromeClient instance =
-        instanceManager.getInstance(instanceId) as WebChromeClient;
-    instance.onProgressChanged(
-      instanceManager.getInstance(webViewInstanceId) as WebView,
-      progress,
+    final WebChromeClient? instance =
+        instanceManager.getInstance(instanceId) as WebChromeClient?;
+    final WebView? webViewInstance =
+        instanceManager.getInstance(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
+      'InstanceManager does not contain an WebChromeClient with instanceId: $instanceId',
     );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+    );
+    instance!.onProgressChanged(webViewInstance!, progress);
   }
 }