+++ packages/shared_preferences/shared_preferences/CHANGELOG.md
@@ -1,6 +1,9 @@
-## NEXT
+## 2.0.9
 
 * Fixes newly enabled analyzer options.
+* Updates example app Android compileSdkVersion to 31.
+
+* Moved Android and iOS implementations to federated packages.
 
 ## 2.0.8