+++ packages/webview_flutter/webview_flutter/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.4.0
 
-* Add an option to set the background color of the webview.
+* Adds an option to set the background color of the webview.
+* Updates example app Android compileSdkVersion to 31.
+* Integration test fixes.
 
 ## 2.3.1